requirements.txt
streamlit
groq>=0.9.0
python-dotenv